Move Gmail callback to API
