+++ callback/config.example.php
@@ -1,5 +1,5 @@
 <?php
 return [
-    'post_to' => 'http://localhost/settings/gmail/callback-shadow',
+    'post_to' => 'http://localhost/api/gmail/callback-shadow',
     'redirect_to' => 'http://localhost/settings/gmail',
 ];
